♻️ refactor: split e2e.spec.ts into granular test files by feature
- Split monolithic e2e.spec.ts into 5 focused test files
- Created vault/initialization.spec.ts for vault auto-init tests
- Created vault/entity-creation.spec.ts for entity creation and search
- Created keyboard-shortcuts/connect-mode.spec.ts for connect mode toggle
- Created keyboard-shortcuts/label-toggle.spec.ts for label visibility
- Created navigation/front-page.spec.ts for welcome screen behavior
- All tests pass successfully (5/5)
